score_video_files.sh: use associative arrays for speed boost
Using a Bash 4.0+ associative array as a hash lookup mechanism to
see if a file path is already in the in-memory file list results
in an absurdly massive performance boost. In testing on a score
list for 15,257 archived video files, the approximate speed boost
was 9447x faster. This number was extrapolated because the faster
version finished in 19 seconds while the slower version was only
done with 3600 of the 15257 files after 306 seconds (5m6s).